+++ src/index.css
@@ -104,11 +104,6 @@ button:focus-visible {
   margin-bottom: 5px;
 }
 
-.table-name-input {
-  margin-left: 15px;
-  height: 25px;
-  width: 155px;
-}
 
 
 @media (prefers-color-scheme: light) {
